+++ src/tabular_trees/sklearn/__init__.py
@@ -1,0 +1,3 @@
+"""Module containing functionality for dealing with sklearn models."""
+
+from . import trees